backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
ccxt>=4.1.0
pandas>=2.1.3
numpy>=1.26.2
requests>=2.31.0
python-dotenv>=1.0.0
